fix: test dummy input

+++ rgcn_kbqa2.py
@@ -128,7 +128,7 @@ class KBQA_RGCN:
         # https://github.com/tkipf/relational-gcn
         # TODO make tensor out of constant
         kb_entities_input = Input(shape=(self.num_entities,), name='entities_input')
-        kb_adjacency_input = [K.variable(kb_relation_adjacency) for kb_relation_adjacency in self.kb_adjacency]
+        kb_adjacency_input = [K.variable(kb_relation_adjacency, dtype='float64') for kb_relation_adjacency in self.kb_adjacency]
 
         # E' - question words embedding
         question_embedding_output = words_embeddings(question_input)
@@ -181,7 +181,7 @@ class KBQA_RGCN:
         # # TODO what happens with 0 representation entities ?
         # kb_entities_labels_word_indices = [self.wordToIndex[entity_label] if entity_label in self.wordToIndex.keys() else 0 for entity_label in kb_entities_labels]
         # print kb_entities_labels_word_indices[:5]
-        kb_entities = np.random.randint(low=1, high=self.num_entities+1, size=(self.num_samples, self.num_entities), dtype='float64')
+        kb_entities = np.random.randint(low=1, high=self.num_entities+1, size=(self.num_samples, self.num_entities))
         # print("Dimensions of the KB entities labels word indices: %s"%str(kb_entities_labels_word_indices.shape))
         # kb_entities = np.array(kb_entities_labels_word_indices * self.num_samples)
         print("Dimensions of the KB entities batches: %s"%str(kb_entities.shape))
